+++ src/app.tsx
@@ -1,6 +1,6 @@
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { HashRouter, Route, Routes } from 'react-router-dom';
 import getStore from './features/store';
 import NavBar from './app/navigation/NavBar';
 import './app.scss';
@@ -10,11 +10,9 @@ import LanguageProvider from './lang/LanguageProvider';
 import 'bootstrap';
 import type { IRoute } from './router';
 
-const basepath = import.meta.env.BASE_URL;
-
 const tools: IRoute[] = [
   {
-    route: `${basepath}fortress`,
+    route: '/fortress',
     title: 'fortress.title',
     description: 'fortress.description',
     page: Fortress,
@@ -27,19 +25,19 @@ export function App() {
     <Provider store={store}>
       <PersistGate loading={null} persistor={persistor}>
         <LanguageProvider>
-          <BrowserRouter>
+          <HashRouter>
             <NavBar tools={tools} />
             <main className="container">
               <div className="bg-light p-5 rounded">
                 <Routes>
-                  <Route path={basepath} element={<Home tools={tools} />} />
+                  <Route path="/" element={<Home tools={tools} />} />
                   {tools.map(({ route, page: Page }) => (
                     <Route key={route} path={route} element={<Page />} />
                   ))}
                 </Routes>
               </div>
             </main>
-          </BrowserRouter>
+          </HashRouter>
         </LanguageProvider>
       </PersistGate>
     </Provider>