+++ src/app/navigation/NavBar.tsx
@@ -18,7 +18,7 @@ const NavBar = ({ tools }: INavBarProps) => {
   return (
     <nav className="navbar navbar-expand-lg navbar-dark fixed-top bg-dark">
       <div className="container-fluid">
-        <Link className="navbar-brand" to={import.meta.env.BASE_URL}>
+        <Link className="navbar-brand" to="/">
           {app.title}
         </Link>
         <button
@@ -35,11 +35,7 @@ const NavBar = ({ tools }: INavBarProps) => {
         <div className="collapse navbar-collapse" id="navbarCollapse">
           <ul className="navbar-nav me-auto mb-2 mb-md-0">
             <li className="nav-item">
-              <Link
-                className="nav-link active"
-                aria-current="page"
-                to={import.meta.env.BASE_URL}
-              >
+              <Link className="nav-link active" aria-current="page" to="/">
                 {app.home}
               </Link>
             </li>